+++ src/App.vue
@@ -2,6 +2,33 @@
   <div id="q-app">
     <div id="transparent-layer" v-if="dim"></div>
     <router-view />
+    <q-modal v-model="showNewUser" ref="newUserModal" content-classes="new-user-modal" no-route-dismiss no-esc-dismiss no-backdrop-dismiss>
+      <div class="row gutter-md">
+        <div class="col-12">
+          <h4>New User</h4>
+          <p>Welcome! Please provide a new password for your account.</p>
+        </div>
+        <div class="col-12">
+          <q-field
+            :error="$v.newPassword.$error"
+            error-label="Passwords must be at least 8 characters"
+          >
+            <q-input v-model="newPassword" type="password" float-label="New Password" @blur="$v.newPassword.$touch" @keyup.enter="setPassword" />
+          </q-field>
+        </div>
+        <div class="col-12">
+          <q-field
+            :error="$v.newPasswordCheck.$error"
+            error-label="Passwords must be identical"
+          >
+            <q-input v-model="newPasswordCheck" type="password" float-label="Retype New Password" @blur="$v.newPasswordCheck.$touch" @keyup.enter="setPassword" />
+          </q-field>
+        </div>
+        <div class="col-12">
+          <q-btn color="primary" @click.native="setPassword">Update</q-btn>
+        </div>
+      </div>
+    </q-modal>
   </div>
 </template>
 
@@ -11,7 +38,10 @@
 import FastClick from 'fastclick'
 FastClick.notNeeded = true
 
+import { required, minLength, sameAs } from 'vuelidate/lib/validators'
+
 var initUser = {
+  newUser: true,
   nqUser: false,
   realUser: false,
   theme: '',
@@ -52,7 +82,19 @@ export default {
   data () {
     return {
       user: initUser,
-      dim: false
+      dim: false,
+      showNewUser: false,
+      newPassword: '',
+      newPasswordCheck: ''
+    }
+  },
+  validations: {
+    newPassword: {
+      required,
+      minLength: minLength(8)
+    },
+    newPasswordCheck: {
+      sameAsPassword: sameAs('newPassword')
     }
   },
   mounted () {
@@ -82,17 +124,49 @@ export default {
         default:
           console.error('invalid theme')
       }
+    },
+    'user.newUser': function (val) {
+      if (val === undefined || val === false) return
+      if (val === true) {
+        this.showNewUser = true
+      }
     }
   },
   methods: {
     init () {
+      this.newPassword = ''
+      this.newPasswordCheck = ''
       this.$firebase.auth.onAuthStateChanged((user) => {
         console.log('auth state changed')
         if (!user) {
           console.log('no user')
+          this.user.theme = 'light'
         } else {
           this.$bindAsObject('user', this.$firebase.user() || initUser)
         }
+      })
+    },
+    setPassword () {
+      this.$v.newPassword.$touch()
+      this.$v.newPasswordCheck.$touch()
+      if (this.$v.newPassword.$error || this.$v.newPasswordCheck.$error) {
+        this.$q.notify.create('Please review password fields')
+      }
+      this.$firebase.auth.currentUser.updatePassword(this.newPassword).then(() => {
+        this.$firebase.user().update({
+          newUser: false
+        })
+        this.$q.notify({
+          type: 'positive',
+          message: 'New password set!',
+          position: 'bottom-left'
+        })
+        this.showNewUser = false
+        this.newPassword = ''
+        this.newPasswordCheck = ''
+      }).catch((error) => {
+        console.error(error)
+        this.$q.notify(error.message)
       })
     }
   }
@@ -175,6 +249,18 @@ export default {
 .active-card {
   z-index: 2;
   position: relative;
+}
+
+.new-user-modal {
+  padding: 30px;
+  width: 100%;
+}
+
+@media screen and (min-width: 1200px) {
+  .new-user-modal {
+    min-width: 500px;
+    width: 500px;
+  }
 }
 
 </style>