+++ src/pages/Settings.vue
@@ -15,12 +15,13 @@
       <div class="col-12" v-if="$root.user.name">
         <q-card>
           <q-card-title>
-            <q-btn color="primary" class="float-right">Edit</q-btn>
+            <q-btn color="secondary" class="float-right" style="margin-left: 10px;" @click.native="authenticate = true">Password</q-btn>
+            <q-btn color="primary" class="float-right" @click.native="openEditUser">Edit</q-btn>
             User Info
           </q-card-title>
           <q-card-main>
             <p><b>Name:</b> <span>{{ $root.user.name.first }} {{ $root.user.name.last }}</span></p>
-            <p><b>Email:</b> <span>{{ $firebase.auth.currentUser.email }}</span></p>
+            <p><b>Email:</b> <span>{{ $root.user.email }}</span></p>
           </q-card-main>
         </q-card>
       </div>
@@ -87,14 +88,95 @@
         </q-card>
       </div>
     </div>
+    <q-modal v-model="editUser" ref="editUserModal" content-classes="edit-settings-modal">
+      <div class="row gutter-md">
+        <div class="col-12">
+          <q-btn
+            color="primary"
+            @click.native="editUser = false; emailEdited = false;"
+            icon="fas fa-times"
+            class="float-right"
+            size="sm"
+          />
+          <h4>Edit User</h4>
+        </div>
+        <div class="col-12">
+          <q-field
+            :error="$v.email.$error"
+            error-label="Must enter a valid email address"
+          >
+            <q-input v-model="email" type="email" float-label="Email" @blur="$v.email.$touch" @input="emailEdited = true" />
+          </q-field>
+        </div>
+        <div class="col-xs-12 col-md-6">
+          <q-input v-model="name.first" float-label="First" />
+        </div>
+        <div class="col-xs-12 col-md-6">
+          <q-input v-model="name.last" float-label="Last" />
+        </div>
+        <div class="col-12">
+          <q-btn color="primary" @click.native="updateUser">Save</q-btn>
+        </div>
+      </div>
+    </q-modal>
+    <q-modal v-model="authenticate" ref="authenticateModal" content-classes="edit-settings-modal">
+      <div class="row gutter-md">
+        <div class="col-12">
+          <q-btn
+            color="primary"
+            @click.native="authenticate = false; emailEdited = false"
+            icon="fas fa-times"
+            class="float-right"
+            size="sm"
+          />
+          <h4>Re-Authenticate</h4>
+        </div>
+        <div class="col-12">
+          <q-input v-model="password" type="password" float-label="Password" @keyup.enter="reauthenticate" />
+        </div>
+        <div class="col-12" v-if="!emailEdited">
+          <q-field
+            :error="$v.newPassword.$error"
+            error-label="Passwords must be at least 8 characters"
+          >
+            <q-input v-model="newPassword" type="password" float-label="New Password" @blur="$v.newPassword.$touch" />
+          </q-field>
+        </div>
+        <div class="col-12" v-if="!emailEdited">
+          <q-field
+            :error="$v.newPasswordCheck.$error"
+            error-label="Passwords must be identical"
+          >
+            <q-input v-model="newPasswordCheck" type="password" float-label="Retype New Password" @blur="$v.newPasswordCheck.$touch" />
+          </q-field>
+        </div>
+        <div class="col-12">
+          <q-btn color="primary" @click.native="reauthenticate">Authenticate</q-btn>
+        </div>
+      </div>
+    </q-modal>
   </q-page>
 </template>
 
 <script>
+import { Notify } from 'quasar'
+import { required, email, sameAs, minLength } from 'vuelidate/lib/validators'
+
 export default {
   // name: 'PageName',
   data () {
     return {
+      name: {
+        first: '',
+        last: ''
+      },
+      email: '',
+      password: '',
+      newPassword: '',
+      newPasswordCheck: '',
+      editUser: false,
+      emailEdited: false,
+      authenticate: false,
       themeOptions: [
         {
           label: 'Light',
@@ -145,41 +227,95 @@ export default {
       ]
     }
   },
+  validations: {
+    email: {
+      required,
+      email
+    },
+    newPassword: {
+      required,
+      minLength: minLength(8)
+    },
+    newPasswordCheck: {
+      sameAsPassword: sameAs('newPassword')
+    }
+  },
   methods: {
+    reauthenticate () {
+      if (!this.emailEdited) {
+        this.$v.newPassword.$touch()
+        this.$v.newPasswordCheck.$touch()
+        if (this.$v.newPassword.$error || this.$v.newPasswordCheck.$error) {
+          this.$q.notify('Please review fields again')
+          return
+        }
+      }
+      const credential = this.$firebase.emailCred(
+        this.$firebase.auth.currentUser.email,
+        this.password
+      )
+      this.$firebase.auth.currentUser.reauthenticateWithCredential(credential).then(() => {
+        if (this.emailEdited) {
+          this.$firebase.auth.currentUser.updateEmail(this.$root.user.email)
+          this.$firebase.user().update({
+            email: this.$root.user.email
+          })
+          this.emailEdited = false
+          Notify.create({
+            type: 'positive',
+            message: 'Email Updated!',
+            position: 'bottom-left'
+          })
+        } else {
+          this.$firebase.auth.currentUser.updatePassword(this.newPassword).then(() => {
+            Notify.create({
+              type: 'positive',
+              message: 'Password Updated!',
+              position: 'bottom-left'
+            })
+          }).catch((error) => {
+            Notify.create({
+              type: 'negative',
+              message: error.message,
+              position: 'bottom-left'
+            })
+          })
+        }
+        this.authenticate = false
+      })
+    },
+    openEditUser () {
+      this.email = this.$root.user.email
+      this.name = {
+        first: this.$root.user.name.first,
+        last: this.$root.user.name.last
+      }
+      this.editUser = true
+    },
+    updateUser () {
+      this.$v.email.$touch()
+      if (this.$v.email.$error) {
+        this.$q.notify('Please review fields again')
+        return
+      }
+      if (this.emailEdited) {
+        this.editUser = false
+        this.authenticate = true
+      }
+      this.$firebase.user().update({
+        name: this.$root.user.name
+      }).then(() => {
+        this.editUser = false
+      })
+    },
     themeChange (val) {
+      // GA - Theme change event
+      this.$ga.event('user', 'theme-change', this.$root.user.theme)
       console.log('change...', val)
       this.$firebase.user().update({
         theme: this.$root.user.theme
       })
     },
-    // translationChange (val) {
-    //   console.log('change...', val)
-    //   this.$firebase.user().update({
-    //     prefs: {
-    //       bibleTranslation: this.$root.user.prefs.bibleTranslation
-    //     }
-    //   })
-    // },
-    // contentChange (val) {
-    //   console.log('val', val)
-    //   var obj = {}
-    //   obj[val] = true
-    //   // this.$firebase.user().update({
-    //   //   prefs: {
-    //   //     contentType: obj
-    //   //   }
-    //   // })
-    // },
-    // mediaChange (val) {
-    //   console.log('valu', val)
-    //   var obj = {}
-    //   obj[val] = true
-    //   // this.$firebase.user().update({
-    //   //   prefs: {
-    //   //     mediaType: obj
-    //   //   }
-    //   // })
-    // },
     prefChange (val) {
       this.$firebase.user().update({
         prefs: this.$root.user.prefs
@@ -190,4 +326,17 @@ export default {
 </script>
 
 <style>
+
+.edit-settings-modal {
+  padding: 30px;
+  width: 100%;
+}
+
+@media screen and (min-width: 1200px) {
+  .edit-settings-modal {
+    min-width: 500px;
+    width: 500px;
+  }
+}
+
 </style>